Refactor RunTimeSettings and Style to reduce code
Refactor RunTimeSettings and Style to reduce code duplication in updateRunTimeSettings, updatePromptConfig, and updateRuntimeStyle functions by introducing a helper function applyIfSet that copies the dereferenced value of src into dst if src is non-nil. This change simplifies the code and improves maintainability while preserving the existing functionality.
