+++ oviewer/oviewer.go
@@ -10,7 +10,6 @@ import (
 	"os"
 	"os/signal"
 	"path/filepath"
-	"regexp"
 	"slices"
 	"strings"
 	"sync"
@@ -199,139 +198,6 @@ func (r RulerType) String() string {
 	}
 }
 
-// RunTimeSettings structure contains the RunTimeSettings of the display.
-// RunTimeSettings contains values that determine the behavior of each document.
-type RunTimeSettings struct {
-	// Name is the name of the view mode.
-	Name string
-	// Converter is the converter name.
-	Converter string
-	// Caption is an additional caption to display after the file name.
-	Caption string
-	// ColumnDelimiterReg is a compiled regular expression of ColumnDelimiter.
-	ColumnDelimiterReg *regexp.Regexp
-	// ColumnDelimiter is a column delimiter.
-	ColumnDelimiter string
-	// SectionDelimiterReg is a section delimiter.
-	SectionDelimiterReg *regexp.Regexp
-	// SectionDelimiter is a section delimiter.
-	SectionDelimiter string
-	// Specified string for jumpTarget.
-	JumpTarget string
-	// MultiColorWords specifies words to color separated by spaces.
-	MultiColorWords []string
-
-	// TabWidth is tab stop num.
-	TabWidth int
-	// Header is number of header lines to be fixed.
-	Header int
-	// VerticalHeader is the number of vertical header lines.
-	VerticalHeader int
-	// HeaderColumn is the number of columns from the left to be fixed.
-	// If 0 is specified, no columns are fixed.
-	HeaderColumn int
-	// SkipLines is the rows to skip.
-	SkipLines int
-	// WatchInterval is the watch interval (seconds).
-	WatchInterval int
-	// MarkStyleWidth is width to apply the style of the marked line.
-	MarkStyleWidth int
-	// SectionStartPosition is a section start position.
-	SectionStartPosition int
-	// SectionHeaderNum is the number of lines in the section header.
-	SectionHeaderNum int
-	// HScrollWidth is the horizontal scroll width.
-	HScrollWidth string
-	// HScrollWidthNum is the horizontal scroll width.
-	HScrollWidthNum int
-	// VScrollLines is the number of lines to scroll with the mouse wheel.
-	VScrollLines int
-	// RulerType is the ruler type (0: none, 1: relative, 2: absolute).
-	RulerType RulerType
-	// AlternateRows alternately style rows.
-	AlternateRows bool
-	// ColumnMode is column mode.
-	ColumnMode bool
-	// ColumnWidth is column width mode.
-	ColumnWidth bool
-	// ColumnRainbow is column rainbow.
-	ColumnRainbow bool
-	// LineNumMode displays line numbers.
-	LineNumMode bool
-	// WrapMode is wrap mode.
-	WrapMode bool
-	// FollowMode is the follow mode.
-	FollowMode bool
-	// FollowAll is a follow mode for all documents.
-	FollowAll bool
-	// FollowSection is a follow mode that uses section instead of line.
-	FollowSection bool
-	// FollowName is the mode to follow files by name.
-	FollowName bool
-	// PlainMode is whether to enable the original character decoration.
-	PlainMode bool
-	// SectionHeader is whether to display the section header.
-	SectionHeader bool
-	// HideOtherSection is whether to hide other sections.
-	HideOtherSection bool
-	// StatusLine is whether to display the status line.
-	StatusLine bool
-
-	// PromptConfig is the prompt configuration.
-	OVPromptConfig
-	// Style is the style of the document.
-	Style Style
-}
-
-// Style structure contains the style settings of the display.
-type Style struct {
-	// ColumnRainbow is the style that applies to the column rainbow color highlight.
-	ColumnRainbow []OVStyle
-	// MultiColorHighlight is the style that applies to the multi color highlight.
-	MultiColorHighlight []OVStyle
-	// Header is the style that applies to the header.
-	Header OVStyle
-	// Body is the style that applies to the body.
-	Body OVStyle
-	// LineNumber is a style that applies line number.
-	LineNumber OVStyle
-	// SearchHighlight is the style that applies to the search highlight.
-	SearchHighlight OVStyle
-	// ColumnHighlight is the style that applies to the column highlight.
-	ColumnHighlight OVStyle
-	// MarkLine is a style that marked line.
-	MarkLine OVStyle
-	// SectionLine is a style that section delimiter line.
-	SectionLine OVStyle
-	// VerticalHeader is a style that applies to the vertical header.
-	VerticalHeader OVStyle
-	// JumpTargetLine is the line that displays the search results.
-	JumpTargetLine OVStyle
-	// Alternate is a style that applies line by line.
-	Alternate OVStyle
-	// Ruler is a style that applies to the ruler.
-	Ruler OVStyle
-	// HeaderBorder is the style that applies to the boundary line of the header.
-	// The boundary line of the header refers to the visual separator between the header and the rest of the content.
-	HeaderBorder OVStyle
-	// SectionHeaderBorder is the style that applies to the boundary line of the section header.
-	// The boundary line of the section header is the line that separates different sections in the header.
-	SectionHeaderBorder OVStyle
-	// VerticalHeaderBorder is the style that applies to the boundary character of the vertical header.
-	// The boundary character of the vertical header refers to the visual separator that delineates the vertical header from the rest of the content.
-	VerticalHeaderBorder OVStyle
-	// LeftStatus is the style that applies to the left status line.
-	LeftStatus OVStyle
-	// RightStatus is the style that applies to the right status line.
-	RightStatus OVStyle
-	// SelectActive is the style that applies to the text being selected (during mouse drag).
-	SelectActive OVStyle
-	// SelectCopied is the style that applies to the text that has been copied to clipboard.
-	SelectCopied OVStyle
-	// PauseLine is the style that applies to the line where follow mode is paused.
-	PauseLine OVStyle
-}
-
 var (
 	// MemoryLimit is a number that limits the chunks to load into memory.
 	MemoryLimit int
@@ -368,18 +234,6 @@ const QuitSmallCountDown = 10
 // MaxWriteLog is the maximum number of lines to output to the log
 // when the debug flag is enabled.
 const MaxWriteLog int = 10
-
-// The name of the converter that can be specified.
-const (
-	convEscaped  string = "es"       // convEscaped processes escape sequence(default).
-	convRaw      string = "raw"      // convRaw is displayed without processing escape sequences as they are.
-	convAlign    string = "align"    // convAlign is aligned in each column.
-	convWordWrap string = "wordwrap" // convWordWrap is wrapped at word boundaries.
-)
-
-const (
-	nameGeneral string = "general"
-)
 
 var (
 	// ErrOutOfRange indicates that value is out of range.
